Collapsed old-style .primitive module and Primitive name (since tests are already failing on this branch following smaller modifications)
